pyproject.toml
[tool.poetry]
name = "trend-tracker"
version = "0.1.0"
description = ""
authors = ["Baptiste Pasquier <pasquierbaptiste@hotmail.fr>"]
readme = "README.md"
packages = [{include = "trend_tracker", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.9.7 || >3.9.7,<4.0"
tweepy = "^4.12.1"
kafka-python = "^2.0.2"
nltk = "^3.8"
contractions = "^0.1.73"
river = "^0.14.0"
streamlit = "^1.16.0"
praw = "^7.6.1"
plotly = "^5.11.0"
wordcloud = "^1.8.2.2"
matplotlib = "^3.6.2"
pymongo = {extras = ["srv"], version = "^4.3.3"}
pymongo-schema = "^0.4.1"
pyyaml = "^6.0"
python-dateutil = "^2.8.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^22.12.0"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pre-commit = "^2.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
